requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
gspread>=6.0.0
google-auth>=2.25.0
python-dotenv>=1.0.0
pyarrow>=14.0.0
slack-sdk>=3.26.0
